+++ src/helper_functions.rs
@@ -1,3 +1,4 @@
+//This function can make a hash out of binary data.
 pub fn fletcher16(data: &[u8]) -> u16 {
     let mut sum1: u16 = 0;
     let mut sum2: u16 = 0;
